+++ R/get_predictions_bayesx.R
@@ -1,0 +1,14 @@
+get_predictions_bayesx <- function(model, fitfram, ...) {
+  prdat <- suppressMessages(
+    stats::predict(
+      object = model,
+      newdata = fitfram,
+      type = "link"
+    ))
+
+  fitfram$predicted <- as.vector(prdat)
+  fitfram$conf.low <- NA
+  fitfram$conf.high <- NA
+
+  fitfram
+}